Use interpolator style
