+++ src/index.js
@@ -1,6 +1,8 @@
 import React from 'react'
 import PropTypes from 'prop-types'
 
+import { linearInterpolator, logInterpolator } from './interpolators'
+
 const getBoundingClientRect = element => {
   const rect = element.getBoundingClientRect()
   return {
@@ -16,7 +18,10 @@ class ReactRanger extends React.Component {
     max: PropTypes.number.isRequired,
     value: PropTypes.oneOfType([PropTypes.number, PropTypes.arrayOf(PropTypes.number)]).isRequired,
     stepSize: PropTypes.number,
-    interpolation: PropTypes.oneOf(['linear', 'logarithmic']),
+    interpolator: PropTypes.shape({
+      getPercentageForValue: PropTypes.func,
+      getValueForClientX: PropTypes.func,
+    }),
     tickSize: PropTypes.number,
     steps: PropTypes.arrayOf(PropTypes.number),
     ticks: PropTypes.arrayOf(PropTypes.number),
@@ -27,7 +32,7 @@ class ReactRanger extends React.Component {
     onRelease: PropTypes.func,
   }
   static defaultProps = {
-    interpolation: 'linear',
+    interpolator: linearInterpolator,
     tickSize: 10,
   }
   state = {
@@ -146,45 +151,13 @@ class ReactRanger extends React.Component {
   }
   getValues = () => (Array.isArray(this.props.value) ? this.props.value : [this.props.value])
   getPercentageForValue = val => {
-    const { min, max, interpolation } = this.props
-
-    if (interpolation === 'logarithmic') {
-      const minSign = Math.sign(min)
-      const maxSign = Math.sign(max)
-
-      if (minSign !== maxSign) {
-        throw new Error(
-          'Error: logarithmic interpolation does not support ranges that cross 0.'
-        )
-      }
-
-      let percent = 100 / (Math.log10(Math.abs(max)) - Math.log10(Math.abs(min))) * (Math.log10(Math.abs(val)) - Math.log10(Math.abs(min)));
-
-      if (minSign < 0) {
-        // negative range, means we need to invert our percent because of the Math.abs above
-        return 100 - percent;
-      }
-      
-      return percent;
-    }
-    return Math.max(0, Math.min(100, ((val - min) / (max - min)) * 100))
+    const { min, max, interpolator } = this.props
+    return interpolator.getPercentageForValue(val, min, max)
   }
   getValueForClientX = clientX => {
-    const { min, max, interpolation } = this.props
+    const { min, max, interpolator } = this.props
     const trackDims = getBoundingClientRect(this.trackEl)
-    const { left, width } = trackDims;
-
-    if (interpolation === 'logarithmic') {
-      let value = clientX - left
-      value *= Math.log10(max) - Math.log10(min);
-      value /= width;
-      value = Math.pow(10, Math.log10(min) + value);
-      return value
-    }
-
-    const percentageValue = (clientX - trackDims.left) / trackDims.width
-    const value = (max - min) * percentageValue
-    return value + min
+    return interpolator.getValueForClientX(clientX, trackDims, min, max)
   }
   getTicks = () => {
     const {
@@ -328,3 +301,4 @@ class ReactRanger extends React.Component {
 }
 
 export default ReactRanger
+export { linearInterpolator, logInterpolator }
