+++ src/interpolators/index.js
@@ -1,0 +1,7 @@
+import linearInterpolator from './linear'
+import logInterpolator from './logarithmic'
+
+export {
+    linearInterpolator,
+    logInterpolator,
+}
